refactor(pipeline): disambiguate the pipeline_validator near-name collision
Resolves the recorded TO-DO collision between
gnn/utils/pipeline_validator.py (pre-execution prerequisite checker;
production import in main.py) and gnn/pipeline/pipeline_validator.py
(runtime integration tester).

- Rename the lower-traffic runtime integration tester to
  gnn/pipeline/pipeline_runtime_validator.py (role-accurate name; one
  test importer, zero production imports at the old path).
- Old import path kept as a compatibility module emitting
  DeprecationWarning and re-exporting PipelineValidator/main with
  identical behavior; contract pinned by
  test_old_import_path_warns_and_reexports (order-independent via
  sys.modules reset).
- Migrate the missed relative-import caller:
  gnn/pipeline/health_check.py (from .pipeline_validator import ...) ->
  canonical path.
- Fix pre-existing collection breakage in
  tests/pipeline/test_pipeline_overall.py: the module-level
  import_module('pipeline') binding (top-level 'pipeline' module does
  not exist; binding unused) interrupted pytest collection of the
  directory — CI's -m 'not pipeline' deselect masked it. Binding
  removed; 551 tests now collect.
- Cross-referencing See-also docstrings updated; test module moved with
  its subject (test_pipeline_runtime_validator.py).

Verification: ruff format/check whole-tree clean; mypy src clean (622
files); path-reference, terminology, docs gates exit 0; targeted suites
94 passed; harness exit 0 (validate_noncanonical_defs=0).
